userRole: reusable comp for Person and SocialLogin
